backend/requirements.txt
flask
werkzeug
opencv-python
numpy
pillow